fix(#345): preserve prompt trailing newline in YAML export + extract highlightYaml
Minor_changes from the iter-1 review: exportNodeAsYaml chose the clip
indicator (`|2`) when a prompt ended in `\n` but never emitted a physical
trailing newline, and clip only preserves a newline that physically ends the
source — so the `|2` vs `|2-` distinction was a no-op and a trailing newline
was dropped on round-trip (export→import→export not idempotent). Emit the
newline for the clip branch; add two regression tests (front has no YAML
parser, so they assert the emitted indicator + trailing-newline shape the
daemon's chomping keys off).

Also fix the react-refresh/only-export-components error the #345 commit
introduced: it turned `highlightYaml` into a named export in PipelineInfoPanel
(a component file). Extract highlightYaml/highlightLine/highlightValue into a
dedicated yamlHighlight.tsx module (the fix the rule itself recommends);
PipelineInfoPanel and ExportNodeYamlModal import from it. Behaviour identical.

tsc clean, vitest 1229 pass, vite build ok. No Rust changes.
