+++ frontend/src/components/PipelineInfoPanel.tsx
@@ -8,6 +8,7 @@ import type { RunState, PipelineDef } from "../types";
 import { isLiveRun } from "../types";
 import { formatDuration, useRunDuration } from "../lib/runDuration";
 import { serializePipeline } from "../stores/editStore";
+import { highlightYaml } from "./yamlHighlight";
 
 export type TabId = "info" | "manager" | "yaml";
 
@@ -340,86 +341,6 @@ function YamlTab({ pipeline, scrollToLine }: { pipeline: PipelineDef | null; scr
       </pre>
     </div>
   );
-}
-
-// Exported for reuse by the node-export modal (#345), which renders the same
-// syntax-highlighted YAML `<pre>` as this panel's YAML tab.
-export function highlightYaml(yaml: string, errorLine?: number): React.ReactNode {
-  const lines = yaml.split("\n");
-  return lines.map((line, i) => {
-    const lineNum = i + 1;
-    const isError = errorLine != null && lineNum === errorLine;
-    const commentIdx = line.indexOf("#");
-    return (
-      <span
-        key={i}
-        className={isError ? "bg-st-failed/20" : undefined}
-        data-line={lineNum}
-      >
-        {commentIdx >= 0 ? (
-          <>
-            {highlightLine(line.slice(0, commentIdx))}
-            <span className="text-fg-4 italic">{line.slice(commentIdx)}</span>
-          </>
-        ) : (
-          highlightLine(line)
-        )}
-        {i < lines.length - 1 ? "\n" : null}
-      </span>
-    );
-  });
-}
-
-function highlightLine(line: string): React.ReactNode {
-  const keyMatch = line.match(/^(\s*-?\s*)([a-zA-Z_][\w]*)\s*:/);
-  if (keyMatch) {
-    const [, indent, key] = keyMatch;
-    const rest = line.slice(keyMatch[0].length);
-    return (
-      <>
-        {indent}
-        <span className="text-acc">{key}</span>
-        <span className="text-fg-4">:</span>
-        {highlightValue(rest)}
-      </>
-    );
-  }
-
-  const listMatch = line.match(/^(\s*-\s+)(.*)/);
-  if (listMatch) {
-    const [, prefix, rest] = listMatch;
-    return (
-      <>
-        <span className="text-fg-4">{prefix}</span>
-        {highlightValue(rest)}
-      </>
-    );
-  }
-
-  return line;
-}
-
-function highlightValue(value: string): React.ReactNode {
-  const trimmed = value.trimStart();
-  const leading = value.slice(0, value.length - trimmed.length);
-
-  if (/^".*"$/.test(trimmed) || /^'.*'$/.test(trimmed)) {
-    return <>{leading}<span className="text-st-await">{trimmed}</span></>;
-  }
-
-  if (/^(true|false|null)$/.test(trimmed)) {
-    return <>{leading}<span className="text-st-running">{trimmed}</span></>;
-  }
-
-  if (/^-?\d+(\.\d+)?$/.test(trimmed)) {
-    return <>{leading}<span className="text-st-done">{trimmed}</span></>;
-  }
-
-  if (/^\{.*\}$/.test(trimmed)) {
-    return <>{leading}<span className="text-fg-3">{trimmed}</span></>;
-  }
-
-  return <>{leading}<span className="text-fg-2">{trimmed}</span></>;
 }
 
 function formatVariableValue(value: unknown): string {
